--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,22 +1,16 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"regexp"
 	"strings"
+
+	jsoniter "github.com/json-iterator/go"
 )
 
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	Email string
 }
 
 type DomainStat map[string]int
@@ -32,35 +26,36 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 type users [100_000]User
 
 func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
+	json := jsoniter.ConfigCompatibleWithStandardLibrary
+	reader := bufio.NewReader(r)
+	for i := 0; ; i++ {
+		line, err1 := reader.ReadBytes('\n')
+		if len(line) > 0 {
+			var user User
+			if err = json.Unmarshal(line, &user); err != nil {
+				return
+			}
+			result[i] = user
+		}
+		if err1 == io.EOF {
 			return
 		}
-		result[i] = user
 	}
-	return
 }
 
 func countDomains(u users, domain string) (DomainStat, error) {
+	domain = "." + domain
+	n := len(domain)
 	result := make(DomainStat)
-
 	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
+		s := len(user.Email) - n
+		if s < 0 {
+			s = 0
 		}
-
+		matched := user.Email[s:] == domain
 		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+			key := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+			result[key]++
 		}
 	}
 	return result, nil