--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -2,7 +2,6 @@ package hw10programoptimization
 
 import (
 	"bufio"
-	"fmt"
 	"io"
 	"strings"
 
@@ -16,46 +15,32 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
+	domain = "." + domain
+	n := len(domain)
 	json := jsoniter.ConfigCompatibleWithStandardLibrary
 	reader := bufio.NewReader(r)
-	for i := 0; ; i++ {
+	result := make(DomainStat)
+	for {
 		line, err1 := reader.ReadBytes('\n')
 		if len(line) > 0 {
 			var user User
-			if err = json.Unmarshal(line, &user); err != nil {
-				return
+			if err := json.Unmarshal(line, &user); err != nil {
+				return nil, err
 			}
-			result[i] = user
+			offset := len(user.Email) - n
+			if offset < 0 {
+				continue
+			}
+			if user.Email[offset:] == domain {
+				key := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+				result[key]++
+			}
 		}
 		if err1 == io.EOF {
-			return
+			break
 		}
-	}
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
-	domain = "." + domain
-	n := len(domain)
-	result := make(DomainStat)
-	for _, user := range u {
-		s := len(user.Email) - n
-		if s < 0 {
-			s = 0
-		}
-		matched := user.Email[s:] == domain
-		if matched {
-			key := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
-			result[key]++
+		if err1 != nil {
+			return nil, err1
 		}
 	}
 	return result, nil